Completed the partner invitation functionality

+++ app/Http/Controllers/InvitePartnerController.php
@@ -13,27 +13,25 @@ use Illuminate\Support\Str;
 class InvitePartnerController extends Controller
 {
     //show a user a form with an email field to invite a new user
-    public function invite(){
+    public function invite()
+    {
         return view('partner.invite');
     }
 
     //process the form submission and send the invite by email
-    public function process(Request $request){
+    public function process(Request $request)
+    {
         //Validating the incoming request data
-        do{
+        do {
             //Generate a random token
             $token = Str::random(16);
 
         }//keep generating a new token if the token already exists
-        while(PartnerInvite::where('token', $token)->first());
+        while (PartnerInvite::where('token', $token)->first());
 
         $currentUser = Auth::user();
         $invitedUser = User::where('email', $request->get('email'));
 
-//        if ($currentUser->partner_email != null ||
-//            $invitedUser->partner_email != null){
-//
-//        }
         //Create a new invite record and store it in the partner_invites table
         $invite = PartnerInvite::create([
             'sender_id' => $currentUser->getAuthIdentifier(),
@@ -50,28 +48,30 @@ class InvitePartnerController extends Controller
     }
 
     //here we'll look up the user by the token sent provided in the URL
-    public function accept($token){
+    public function accept($token)
+    {
         //Handling if the invite doesn't exist
-        if(!$invite = PartnerInvite::where('token', $token)->first()){
+        if (!$invite = PartnerInvite::where('token', $token)->first()) {
             //Abort the task
             abort(404);
         }
+        if ($invitedUser = User::where('email', $invite->partner_email)->first()) {
+            //Finding both users
+            $currentUser = User::where('id', $invite->sender_id)->first();
 
-        //Finding both users
-        $currentUser = User::where('id', $invite->sender_id)->first();
-        $invitedUser = User::where('email', $invite->partner_email)->first();
+            //Assigning both users partner emails to each other
+            $currentUser->partner_email = $invitedUser->email;
+            $invitedUser->partner_email = $currentUser->email;
 
-        //Assigning both users partner emails to each other
-        $currentUser->partner_email = $invitedUser->partner_email;
-        $invitedUser->partner_email = $currentUser->partner_email;
+            $currentUser->update(['partner_email' => $invitedUser->partner_email]);
+            $invitedUser->save();
 
-        $currentUser->update(['partner_email' => $invitedUser->partner_email]);
-        $invitedUser->save();
+            //Delete the invite record so it is not reused
+            $invite->delete();
 
-        //Delete the invite record so it is not reused
-        $invite->delete();
-
-        //Here we will put the actions that will happen after the invitation is done successfully but for now we will prove it worked
-        return 'Good job! ' . $invitedUser->name . ' Invite accepted!';
+            //Here we will put the actions that will happen after the invitation is done successfully but for now we will prove it worked
+            return 'Good job! ' . $invitedUser->name . ' Invite accepted!';
+        }
+        return "Invited user account not found!";
     }
 }
